saveGame moved to create page
