+++ components/puzzle/GameCreator/ColumnDecorator.js
@@ -11,7 +11,7 @@ const defaultBorders = [
     {borderPosition:'L', buttonTitle:'Border Left'},
 ]
 
-export default function ColumnDecorator(props){
+export default function ColumnDecorator(){
     const createModeThomas = useSelector(state=>state.createGame.thomas);
     const createModeWolf = useSelector(state=>state.createGame.wolf);
     const selectedColumns = useSelector(state=>state.createGame.selectedColumns);
@@ -48,8 +48,6 @@ export default function ColumnDecorator(props){
 
                 {selectedColumns.length>0 && <button onClick={onClickUnSelectColumns} className='btn btn-primary m-1 mt-3'>unselect columns</button>}
                 {showUnsetThomasWolfButton && <button onClick={onClickUnSetThomasWolf} className='btn btn-warning m-1 mt-3'>Unset Thomas & Wolf</button>}
-
-                <button onClick={props.onClickSavePuzzleButton} className='btn btn-dark m-1'>Save Puzzle</button>
                 
                 <ButtonSetThomasWolf setType='wolf' btnClass='btn-danger' title='Wolf' />
                 <ButtonSetThomasWolf setType='thomas' btnClass='btn-success' title='Thomas' />
